[FE] refactor : Snackbar 리팩토링
- 변경된 이벤트 핸들러 컨벤션 적용
- defaultProps 제거
- atom type 위치 수정
