+++ frontend/src/common/components/SnackbarProvider/index.tsx
@@ -1,23 +1,23 @@
+import { snackbarStackWithActive } from 'common/recoil/snackbar';
+
 import { useRecoilValue } from 'recoil';
 
 import useSnackbar from 'common/hooks/useSnackbar';
 
 import Snackbar from 'common/components/Snackbar';
 
-import { snackbarStackWithActive } from 'common/recoil/snackbar';
-
 function SnackbarProvider() {
   const snackbarActivityStack = useRecoilValue(snackbarStackWithActive);
   const snackbar = useSnackbar();
 
-  const onSnackbarDisappear = (snackbarId: number) => () => {
+  const handleRemoveSnackbar = (snackbarId: number) => () => {
     snackbar.remove(snackbarId);
   };
 
   return (
     <div id="snackbar-container">
       {snackbarActivityStack.map(({ key, ...rest }) => (
-        <Snackbar key={key} {...rest} onDisappear={onSnackbarDisappear(key)} />
+        <Snackbar key={key} {...rest} onDisappear={handleRemoveSnackbar(key)} />
       ))}
     </div>
   );
