--- frontend/src/common/types/index.ts
@@ -1,10 +1,0 @@
-import { SnackbarProps } from 'common/components/Snackbar';
-
-interface snackbarStackAtom extends SnackbarProps {
-  key: number;
-}
-
-export interface SnackbarAtom {
-  uniqueKey: number;
-  stack: snackbarStackAtom[];
-}
